Eliminate overhead of muting child
Some juggling with file descriptors allows us to avoid wasting time in the child
between fork and exec to dup /dev/null to stdout and stderr. Since we only write
to stderr (and basically only in case of error before the final timing report),
we can delay restoring it until it is really needed, thus also avoiding any
additional overhead between fork and wait in the parent.