Reverse behavior of -k and improve early stopping
Force kill child if it stops only if -k is given, not by default.
This is a more reasonable/desirable default behavior.

Improve early stopping forcibly killing *only* the child process
in the SIGINT handler (and only if it's still running), instead of
killing the entire process group. We can still gracefully handle the
situation and return the real child exit status in case of a single
run with a few additional checks.